backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-Login==0.6.3
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
SQLAlchemy==2.0.21
Mako==1.2.4
python-dotenv==1.0.0
Jinja2==3.1.2
Werkzeug==2.3.7
scikit-learn==1.3.0
pandas==2.0.3
numpy==1.24.3
nltk==3.8.1
spacy==3.6.1
pdfplumber==0.9.0
python-docx==0.8.11
PyPDF2==3.0.1
requests==2.31.0
urllib3==2.0.4
six==1.16.0
python-dateutil==2.8.2
pytz==2023.3
scipy==1.11.2
tqdm==4.66.1
joblib==1.3.2
certifi==2023.7.22
charset-normalizer==3.2.0
idna==3.4

typing-extensions==4.7.1

# Additional dependencies for NLP and AI
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.6.0/en_core_web_sm-3.6.0.tar.gz#egg=en_core_web_sm
